getTemplateWithAssociatedUsers
