+++ lib/templates.ts
@@ -39,46 +39,6 @@ export type UpdateTemplateCommand = {
   notificationsInterval?: NotificationsInterval;
 };
 
-export async function getTemplateWithAssociatedUsers(formID: string): Promise<{
-  formRecord: FormRecord;
-  users: { id: string; name: string | null; email: string }[];
-} | null> {
-  const { user } = await authorization.canViewForm(formID).catch((e) => {
-    logEvent(
-      e.user.id,
-      { type: "Form", id: formID },
-      "AccessDenied",
-      AuditLogAccessDeniedDetails.AccessDenied_AttemptToListAssignedUsers
-    );
-    throw e;
-  });
-  const templateWithAssociatedUsers = await prisma.template
-    .findUnique({
-      where: {
-        id: formID,
-      },
-      include: {
-        deliveryOption: true,
-        users: {
-          select: {
-            id: true,
-            name: true,
-            email: true,
-          },
-        },
-      },
-    })
-    .catch((e) => prismaErrors(e, null));
-
-  if (!templateWithAssociatedUsers) return null;
-
-  logEvent(user.id, { type: "Form", id: formID }, "ReadForm", AuditLogDetails.RetrieveFormUsers);
-  return {
-    formRecord: _parseTemplate(templateWithAssociatedUsers),
-    users: templateWithAssociatedUsers.users,
-  };
-}
-
 /**
  * Update a form template
  * @param template A Form Record containing updated information
